requirements.txt
modal
numpy
pandas
ccxt
# smartmoneyconcepts  <-- Removed unused
google-genai
google-generativeai
python-telegram-bot
python-dotenv
requests
fastapi
uvicorn
mplfinance
Pillow
supabase
psycopg2-binary
